+++ pandas_bokeh/plot.py
@@ -18,17 +18,7 @@ from bokeh.models import (
     LogColorMapper,
     CategoricalColorMapper,
     ColorBar,
-    BasicTicker,
-
     FuncTickFormatter,
-    NumeralTickFormatter,
-    BasicTickFormatter,
-    PrintfTickFormatter,
-    CategoricalTickFormatter,
-    DatetimeTickFormatter,
-    MercatorTickFormatter,
-    LogTickFormatter,
-    TickFormatter,
     WheelZoomTool,
 )
 from bokeh.models.tickers import FixedTicker
@@ -43,7 +33,7 @@ from bokeh.events import Tap
 from pandas.plotting._core import BasePlotMethods
 
 from .base import show, embedded_html
-from .utils import get_tick_formatter
+from .geoplot import geoplot
 
 
 
@@ -1064,7 +1054,6 @@ def scatterplot(
     x_axis_type,
     xlabelname,
     ylabelname,
-    colorbar_tick_format=None,
     **kwargs
 ):
     """Adds a scatterplot to figure p for each data_col."""
@@ -1122,17 +1111,10 @@ def scatterplot(
             # Define Colorbar:
             colorbar_options = {
                 "color_mapper": colormapper,
-                "label_standoff": 12,
+                "label_standoff": 0,
                 "border_line_color": None,
                 "location": (0, 0),
-                "ticker": BasicTicker()
             }
-
-            if colorbar_tick_format:
-                colorbar_options["formatter"] = get_tick_formatter(colorbar_tick_format)
-
-            print(colorbar_options)
-
             colorbar = ColorBar(**colorbar_options)
             p.add_layout(colorbar, "right")
